인기추천 디자인, 마이페이지 사진편집 후 바로 적용

+++ frontend/src/components/popup/recommend/PopupRecommendComponent.js
@@ -5,29 +5,47 @@ import {
   postRecommend,
   deleteRecommend,
 } from "../../../api/popupRecommendApi";
+import { useStyleSheetContext } from "styled-components/dist/models/StyleSheetManager";
 
+//팝업 전용 추천 컴포넌트
 const PopupRecommendComponent = ({ popupId }) => {
+  const userId = useStyleSheetContext((state) => state.auth?.user?.id);
+
   const [isRecommended, setIsRecommended] = useState(false);
   const [Loading, setLoading] = useState(false);
 
   useEffect(() => {
+    if (!userId) return;
     async function fetchStatus() {
-      const data = await getIsRecommended(popupId);
-      setIsRecommended(data.isRecommended);
+      try {
+        const data = await getIsRecommended(popupId);
+        setIsRecommended(data.isRecommended);
+      } catch (e) {
+        console.eeror("추천 가져오기 실패", e);
+      }
     }
     fetchStatus();
   }, [popupId]);
 
   const handleClick = async () => {
-    setLoading(true);
-    if (isRecommended) {
-      await deleteRecommend(popupId);
-      setIsRecommended(false);
-    } else {
-      await postRecommend(popupId);
-      setIsRecommended(true);
+    if (!userId) {
+      alert("로그인 후 추천 가능합니다");
+      return;
     }
-    setLoading(false);
+    setLoading(true);
+    try {
+      if (isRecommended) {
+        await deleteRecommend(popupId);
+        setIsRecommended(false);
+      } else {
+        await postRecommend(popupId);
+        setIsRecommended(true);
+      }
+    } catch (e) {
+      alert("추천 오류 발생", e);
+    } finally {
+      setLoading(false);
+    }
   };
 
   return (
